+++ frontend/src/index.js
@@ -6,6 +6,7 @@ import SingleNote from './pages/notes/[id]'
 import SingleProduct from './pages/products/[id]'
 import CartPage from './pages/CartPage'
 import LoginPage from './pages/LoginPage/LoginPage'
+import SignUpPage from './pages/SignUpPage/SignUpPage'
 
 import 'bootstrap/dist/css/bootstrap.min.css'
 import { BudgetsProvider } from './context/BudgetsContext'
@@ -29,6 +30,7 @@ ReactDOM.render(
               <Route path='/' element={<App />}></Route>
               <Route path='/test' element={<MainPage />}></Route>
               <Route path='/login' element={<LoginPage />}></Route>
+              <Route path='/signup' element={<SignUpPage />}></Route>
               <Route path='/cart' element={<CartPage />}></Route>
               <Route path='/notes/:id' element={<SingleNote />}></Route>
               <Route path='/products/:id' element={<SingleProduct />}></Route>