+++ frontend/src/store.js
@@ -5,6 +5,7 @@ import userReducer from './features/user/userSlice'
 import productsReducer from './features/product/productSlice'
 import notesReducer from './features/note/noteSlice'
 import cartReducer from './features/cart/cartSlice'
+import messageReducer from './features/message/messageSlice'
 
 export const store = configureStore({
   reducer: {
@@ -12,6 +13,7 @@ export const store = configureStore({
     products: productsReducer,
     notes: notesReducer,
     cart: cartReducer,
+    message: messageReducer,
   },
   middleware: [thunk],
 })